Implemented memory-mapped registers
The keyboard status and keyboard data registers are read
by reading from memory

Since the keyboard should be checked when checking the
status register, we cannot simply read from the memory array
anymore

So the mem_read getter was written to include the checking
of the keyboard, and updating of the memory mapped registers
to the reading of memory

Likewise the mem_write setter prevents direct modification of
the memory mapped registers (as this should only be done in
the side effects of reading from the status register)

Note that if the user wants to read the keyboard data register,
they must first read the status register to check if there is
a key being pressed. If there is, the act of reading the status
register will load the key being pressed into the data register.

The check_key function to check if a key is being pressed
has not been implemented yet, and always returns false

Existing memory accesses have been changed to use the getters
and setters